Migrate Task Added events
